Entrega Final

+++ Entrega/src/controllers/products.controller.js
@@ -3,13 +3,14 @@ const productsService = require('../services/services').productsService
 const CustomError = require("../utils/CustomError/CustomError")
 const { createProductErrorInfo } = require("../utils/CustomError/info")
 const { Errors } = require("../utils/CustomError/Errors")
-
+const { userModel } = require('../models/user.model')
+const {sendMail} = require('../utils/sendmail')
 class ProductsController {
     getProducts = async (req, res) => {
         try {
             const pageNumber = parseInt(req.query.page) || 1
             const limitNumber = parseInt(req.query.limit) || 10
-            const { query, sort } = req.query
+            const { query, sort, user } = req.query
             let prevLink = null
             let nextLink = null
             let mquery = {}
@@ -24,7 +25,7 @@ class ProductsController {
             if (sort === 'desc') {
                 sortType = { price: -1 }
             }
-            let products = await productsService.get(mquery, limitNumber, pageNumber, sortType)
+            let products = await productsService.get(mquery, limitNumber, pageNumber, sortType, user)
 
             const { docs, totalDocs, limit, totalPages, page, hasNextPage, hasPrevPage, nextPage, prevPage } = products
             if (prevPage !== null) {
@@ -68,8 +69,8 @@ class ProductsController {
 
     createProduct = async (req, res, next) => {
         try {
-            const { title, description, code, price, status, stock, category, thumbnail, featured } = req.body
-            if (!title || !description || !code || !price || !status || !stock || !category || !thumbnail) {
+            const { title, description, code, price, status, stock, category, thumbnail, featured, owner } = req.body
+            if (!title || !description || !code || !price || !status || !stock || !category || !thumbnail || !owner) {
                 CustomError.createError({
                     name: 'Missing Fields Error',
                     cause: createProductErrorInfo(req.body),
@@ -87,7 +88,8 @@ class ProductsController {
                 stock,
                 category,
                 thumbnail,
-                featured
+                featured,
+                owner
             }
             await productsService.create(product)
             res.status(200).send({
@@ -130,7 +132,42 @@ class ProductsController {
     deleteProduct = async (req, res) => {
         try {
             const { pid } = req.params
-            let result = await productsService.delete(pid)
+            const { owner } = req.body
+            let user =  await userModel.findOne({username: owner})
+            let product = await productsService.getById(pid)
+           if (user.role === 'premium') {
+                let to = user.email
+                let subject = 'Uno de sus productos ha sido eliminado'
+                let html = `<!DOCTYPE html>
+                <html lang="es">
+                <head>
+                    <meta charset="UTF-8">
+                    <title>Cuenta Inhabilitada por Inactividad</title>
+                </head>
+                <body>
+                    <table width="100%" cellspacing="0" cellpadding="0" border="0">
+                        <tr>
+                            <td align="center" style="background-color: #f5f5f5; padding: 20px;">
+                                <table width="600" cellspacing="0" cellpadding="0" border="0" style="background-color: #ffffff; border-radius: 5px; box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);">
+                                    <tr>
+                                        <td align="center" style="padding: 40px;">
+                                            <h1 style="color: #333333; font-family: Arial, sans-serif;">Uno de sus productos ha sido eliminado</h1>
+                                            <p style="color: #666666; font-family: Arial, sans-serif; font-size: 16px;">Estimado usuario,</p>
+                                            <p style="color: #666666; font-family: Arial, sans-serif; font-size: 16px;">Le informamos que el producto ${product.code} ha sido eliminado de Sneakers.</p>
+                                            <p style="color: #666666; font-family: Arial, sans-serif; font-size: 16px;">Gracias por utilizar nuestros servicios.</p>
+                                            <p style="color: #666666; font-family: Arial, sans-serif; font-size: 16px;">Atentamente,</p>
+                                            <p style="color: #666666; font-family: Arial, sans-serif; font-size: 16px;">Sneakers</p>
+                                        </td>
+                                    </tr>
+                                </table>
+                            </td>
+                        </tr>
+                    </table>
+                </body>
+                </html>`
+                await sendMail(to, subject, html)
+            }
+            let result = await productsService.delete(pid) 
             res.status(200).send({
                 status: 'success',
                 payload: result
